requirements.txt
python-dotenv
pandas
seaborn
numpy
certifi
pymongo
pymongo[srv]==3.12
scikit-learn
#-e .